pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cmc-cs040-preslie"
version = "0.1.6"
description = "LLM-based chat/document summarization project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq",
    "python-dotenv",
]

[project.scripts]
docsum-chat = "cmc_cs040_preslie.chat:main"

[tool.setuptools.packages.find]
include = ["cmc_cs040_preslie", "cmc_cs040_preslie.*"]